feat: new charts folder
